% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prj_creation.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(
  db_path,
  db_name,
  prj_name,
  scenarios,
  query_path,
  queries_file = "queries_rfasst_nonCO2.xml"
)
}
\arguments{
\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}

\item{query_path:}{path to the queries file}

\item{query_file:}{queries file name}
}
\value{
dataframe with the data from the query
}
\description{
Add nonCO2 large queries
}
