% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prj_creation.R
\name{fill_queries}
\alias{fill_queries}
\title{fill_queries}
\usage{
fill_queries(
  prj,
  db_path,
  db_name,
  prj_name,
  scenarios,
  query_path,
  query_file = "queries_rfasst_nonCO2.xml"
)
}
\arguments{
\item{prj:}{current project}

\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}

\item{query_path:}{path to the queries file}

\item{query_file:}{queries file name}
}
\value{
prj containing the nonCO2 emissions by sector
}
\description{
Create a folder to save the datasets and file, in case it does not exist
}
