% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_set_globals.R, R/set_globals.R
\name{get_globalColorScales}
\alias{get_globalColorScales}
\title{get_globalColorScales}
\usage{
get_globalColorScales()

get_globalColorScales()
}
\description{


Global vars for scale colors
}
\details{

}
