% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfasstui.R
\docType{package}
\name{rfasstui}
\alias{rfasstui-package}
\alias{rfasstui}
\title{rfasstui}
\description{
\itemize{
\item Github: https://github.com/bc3LC/rfasstui
\item Webpage: https://bc3lc.github.io/rfasstui/}
}
\author{
\strong{Maintainer}: Clàudia Rodés-Bachs \email{claudia.rodes@bc3research.org}

Authors:
\itemize{
  \item Jon Sampedro \email{jon.sampedro@bc3research.org}
}

}
